# api/docs/swagger.yaml
basePath: /services
definitions:
  api_internal_models.OcUser:
    properties:
      created_at:
        type: string
      description:
        type: string
      expire_at:
        type: string
      group:
        type: string
      is_locked:
        type: boolean
      is_online:
        type: boolean
      password:
        type: string
      rx:
        description: in bytes
        type: integer
      traffic_size:
        description: in GiB  >> x * 1024 ** 3
        type: integer
      traffic_type:
        type: integer
      tx:
        description: in bytes
        type: integer
      uid:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  api_internal_models.OcUserActivity:
    properties:
      created_at:
        type: string
      log:
        type: string
    type: object
...
  api_internal_repository.Statistics:
    properties:
      date:
        type: string
      sum_rx:
        type: number
      sum_tx:
        type: number
    type: object
...
  api_internal_routes_middlewares.Unauthorized:
    properties:
      error:
        type: string
    type: object
  api_pkg_utils.ErrorResponse:
    properties:
      error:
        type: string
    required:
    - error
    type: object
  api_pkg_utils.ResponsePagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_records:
        type: integer
    type: object
...
  internal_services_oc_user.OcservUserCreateOrUpdateRequest:
    properties:
      expire_at:
        type: string
      group:
        type: string
      password:
        maxLength: 16
        minLength: 1
        type: string
      traffic_size:
        type: integer
      traffic_type:
        type: integer
      username:
        maxLength: 16
        minLength: 3
        type: string
    required:
    - expire_at
    - group
    - password
    - traffic_type
    - username
    type: object
  internal_services_oc_user.OcservUserLockRequest:
    properties:
      lock:
        type: boolean
    required:
    - lock
    type: object
  internal_services_oc_user.OcservUsersResponse:
    properties:
      meta:
        $ref: '#/definitions/api_pkg_utils.ResponsePagination'
      oc_users:
        items:
          $ref: '#/definitions/api_internal_models.OcUser'
        type: array
    type: object
...
info:
  contact: {}
  description: This is a sample Ocserv User management Api server.
  title: Ocserv User management Example Api
  version: "1.0"
paths:
...
  /api/v1/ocserv/users:
    get:
      consumes:
      - application/json
      description: List of Ocserv Users with pagination
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Page number, starting from 1
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Field to order by
        in: query
        name: pager
        type: string
      - description: Sort order, either ASC or DESC
        enum:
        - ASC
        - DESC
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_services_oc_user.OcservUsersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: List of Ocserv Users
      tags:
      - Ocserv Users
    post:
      consumes:
      - application/json
      description: Create Ocserv User
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Create Ocserv User Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_services_oc_user.OcservUserCreateOrUpdateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: Create Ocserv User
      tags:
      - Ocserv Users
  /api/v1/ocserv/users/:uid:
    delete:
      consumes:
      - application/json
      description: Delete Ocserv User
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ocserv User UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: Delete Ocserv User
      tags:
      - Ocserv Users
    get:
      consumes:
      - application/json
      description: Retrieve Ocserv User by given uid
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ocserv User UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_internal_models.OcUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: Retrieve Ocserv User
      tags:
      - Ocserv Users
    put:
      consumes:
      - application/json
      description: Update Ocserv User
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ocserv User UID
        in: path
        name: uid
        required: true
        type: string
      - description: Update Ocserv User Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_services_oc_user.OcservUserCreateOrUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: Update Ocserv User
      tags:
      - Ocserv Users
  /api/v1/ocserv/users/:uid/activities:
    get:
      consumes:
      - application/json
      description: Activities for Ocserv User by given user UID and Date
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ocserv User UID
        in: path
        name: uid
        required: true
        type: string
      - description: Start date in format YYYY-MM-DD
        in: query
        name: start
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api_internal_models.OcUserActivity'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: Activities for Ocserv User
      tags:
      - Ocserv Users
  /api/v1/ocserv/users/:uid/disconnect:
    post:
      consumes:
      - application/json
      description: Disconnect Ocserv User
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ocserv User UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: Disconnect Ocserv User
      tags:
      - Ocserv Users
  /api/v1/ocserv/users/:uid/lock:
    post:
      consumes:
      - application/json
      description: Lock or Unlock Ocserv User
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ocserv User UID
        in: path
        name: uid
        required: true
        type: string
      - description: Update Ocserv User Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_services_oc_user.OcservUserLockRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: Lock or Unlock Ocserv User
      tags:
      - Ocserv Users
  /api/v1/ocserv/users/:uid/statistics:
    get:
      consumes:
      - application/json
      description: Statistics for Ocserv User by given user UID
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ocserv User UID
        in: path
        name: uid
        required: true
        type: string
      - description: Start date in format YYYY-MM-DD, null=time.Now()
        in: query
        name: start
        type: string
      - description: End date in format YYYY-MM-DD, null=time.Now().AddDate(0, 1,
          0) 1 month
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api_internal_repository.Statistics'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: Statistics for Ocserv User
      tags:
      - Ocserv Users
...
swagger: "2.0"